Fix for main window scrolling when clicking inside shell window.
There is a hidden input element for capturing the keystrokes. When you
type, an event captures the key strokes, sends them to server, the server
echos, yada, yada.

The problem is that this hidden input field is in upper left corner of the
main window (not inside the iframe). Thats not really possible, but that
seems to be where the browser thinks it is cause the style sheet for
shellinabox puts it at -10000. Why? Cause you cannot give focus to a
hidden (display: none) element. Anyway, when you click anywhere inside
the console window, an event captures that, and that hidden input field
gets the focus. The browser wants the focused element to be visible, so it
scrolls the entire window so that 0,0 (the hidden element) is visible.

I fixed this by adding a new style for the input field so that I can treat
that one separate from the rest of the hidden div. This new style puts it
a 0,0 in the iframe.